+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Updated Grid size on medium screens to 7
 * Disabled back button on Page 1 in TablePagination
+* Updated the address of a miner in LatestBlocks
 
 ## v0.1.0